requirements.txt
beautifulsoup4==4.12.2
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
Requests==2.31.0
streamlit==1.24.0
tensorflow==2.12.0
